therapist_chat/requirements.txt
asgiref==3.7.2
blinker==1.6.2
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
colorama==0.4.6
distlib==0.3.7
Django==4.2.7
djangorestframework==3.15.2
filelock==3.12.3
filetype==1.2.0
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msgpack==1.0.8
platformdirs==3.10.0
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.7
sqlparse==0.4.4
supervisor==4.2.5
tzdata==2023.3
ua-parser==0.18.0
virtualenv==20.24.5
Werkzeug==2.3.7
